Store all versions

+++ indexchildren.py
@@ -6,7 +6,10 @@ from pycontainers import compressedfile, hashtable
 
 def StoreFactoryCreate(fina, maskBits = 26, maxCachedPages = 50):
 	try:
-		os.unlink(fina)
+		if os.path.exists(fina):
+			print "Delete old", fina
+			os.unlink(fina)
+			print "done"
 	except:
 		pass
 
@@ -31,31 +34,37 @@ class TagIndex(object):
 	def __del__(self):
 		print "Flushing"
 
-	def AddParent(self, childType, childRef, parentType, parentRef):
+	def AddParent(self, childType, childRef, parentType, parentRef, parentVer):
 		#print childType, childRef, parentType, parentRef
 
 		if childType == "node":
 			if childRef not in self.nodeParentStore:
-				tmp = []
+				tmp = {}
 			else:
 				tmp = self.nodeParentStore[childRef]
-			tmp.append([parentType[0], parentRef])
+			if parentVer not in tmp:
+				tmp[parentVer] = []
+			tmp[parentVer].append([parentType[0], parentRef])
 			self.nodeParentStore[childRef] = tmp
 			
 		if childType == "way":
 			if childRef not in self.wayParentStore:
-				tmp = []
+				tmp = {}
 			else:
 				tmp = self.wayParentStore[childRef]
-			tmp.append([parentType[0], parentRef])
+			if parentVer not in tmp:
+				tmp[parentVer] = []
+			tmp[parentVer].append([parentType[0], parentRef])
 			self.wayParentStore[childRef] = tmp
 
 		if childType == "relation":
 			if childRef not in self.relationParentStore:
-				tmp = []
+				tmp = {}
 			else:
 				tmp = self.relationParentStore[childRef]
-			tmp.append([parentType[0], parentRef])
+			if parentVer not in tmp:
+				tmp[parentVer] = []
+			tmp[parentVer].append([parentType[0], parentRef])
 			self.relationParentStore[childRef] = tmp
 
 	def TagLimitCallback(self, name, depth, attr, childTags, childMembers):
@@ -70,14 +79,21 @@ class TagIndex(object):
 
 		self.objs += 1
 
+		if self.objs % 100000 == 0:
+			print "Flushing"
+			self.nodeParentStore.flush()
+			self.wayParentStore.flush()
+			self.relationParentStore.flush()
+
 		if name == "node":
 			self.nodes += 1
 
 		if name == "way":
 			#print name, childTags, childMembers
 			objId = int(attr['id'])
+			version = int(attr['version'])
 			for chType, chRef, chRole in childMembers:
-				self.AddParent(chType, chRef, name, objId)
+				self.AddParent(chType, chRef, name, objId, version)
 				
 			self.ways += 1
 
@@ -85,8 +101,9 @@ class TagIndex(object):
 			#print name, childTags, childMembers
 			self.relations += 1
 			objId = int(attr['id'])
+			version = int(attr['version'])
 			for chType, chRef, chRole in childMembers:
-				self.AddParent(chType, chRef, name, objId)
+				self.AddParent(chType, chRef, name, objId, version)
 
 if __name__ == "__main__":
 	if len(sys.argv) < 2:
@@ -100,8 +117,8 @@ if __name__ == "__main__":
 	infi = bz2.BZ2File(sys.argv[1])
 
 	nodeParentStore, compFiN = StoreFactoryCreate(sys.argv[2]+".node", 26, 5000)
-	wayParentStore, compFiW = StoreFactoryCreate(sys.argv[2]+".way", 26, 5000)
-	relationParentStore, compFiR = StoreFactoryCreate(sys.argv[2]+".relation", 26, 5000)
+	wayParentStore, compFiW = StoreFactoryCreate(sys.argv[2]+".way", 26, 1000)
+	relationParentStore, compFiR = StoreFactoryCreate(sys.argv[2]+".relation", 26, 1000)
 
 	tagIndex = TagIndex(nodeParentStore, wayParentStore, relationParentStore)
